feat: 增强寄存器管理 - 在 RegisterAllocator 中添加 slot_peak 以跟踪分配的最大寄存器槽位数，提供 peakSlots 方法供外部查询。 - 在 Chunk 结构中引入 max_register_slots 字段，记录编译期观察到的寄存器槽位数，供虚拟机校验使用。 - 更新 ObjFunction 以使用 uint16_t 类型的 max_registers -为语言核增加模块级readme文档。

+++ include/niki/l0_core/vm/object.hpp
@@ -78,7 +78,8 @@ struct ObjFunction {
     Object object_header; // 补齐对象头，使其参与 VM 的 C-style 多态
     uint32_t name_id;
     uint8_t arity; // 参数个数
-    uint8_t max_registers;
+    /// 本函数体使用的寄存器槽位上界（比最大寄存器下标多 1）；供 VM 与 OP_CALL 栈安全检查。
+    uint16_t max_registers;
     niki::Chunk chunk; // 属于这个函数的独立字节码块
 };
